feat(uploads): replace apollo-upload-server v5 with graphql-upload v8
