+++ packages/apollo-server-core/package.json
@@ -25,7 +25,6 @@
   },
   "dependencies": {
     "@apollographql/apollo-tools": "^0.2.6",
-    "@apollographql/apollo-upload-server": "^5.0.3",
     "@apollographql/graphql-playground-html": "^1.6.6",
     "@types/ws": "^6.0.0",
     "apollo-cache-control": "file:../apollo-cache-control",
@@ -40,6 +39,7 @@
     "graphql-subscriptions": "^1.0.0",
     "graphql-tag": "^2.9.2",
     "graphql-tools": "^4.0.0",
+    "graphql-upload": "^8.0.0",
     "json-stable-stringify": "^1.0.1",
     "lodash": "^4.17.10",
     "subscriptions-transport-ws": "^0.9.11",
