+++ packages/apollo-server-core/src/ApolloServer.ts
@@ -249,9 +249,7 @@ export class ApolloServerBase {
       );
 
       if (this.uploadsConfig) {
-        const {
-          GraphQLUpload,
-        } = require('@apollographql/apollo-upload-server');
+        const { GraphQLUpload } = require('graphql-upload');
         if (resolvers && !resolvers.Upload) {
           resolvers.Upload = GraphQLUpload;
         }
