+++ packages/apollo-server-express/package.json
@@ -26,7 +26,6 @@
     "node": ">=6"
   },
   "dependencies": {
-    "@apollographql/apollo-upload-server": "^5.0.3",
     "@apollographql/graphql-playground-html": "^1.6.6",
     "@types/accepts": "^1.3.5",
     "@types/body-parser": "1.17.0",
@@ -38,6 +37,7 @@
     "cors": "^2.8.4",
     "graphql-subscriptions": "^1.0.0",
     "graphql-tools": "^4.0.0",
+    "graphql-upload": "^8.0.0",
     "type-is": "^1.6.16"
   },
   "devDependencies": {
