+++ packages/apollo-server-hapi/package.json
@@ -25,13 +25,13 @@
     "node": ">=8"
   },
   "dependencies": {
-    "@apollographql/apollo-upload-server": "^5.0.3",
     "@apollographql/graphql-playground-html": "^1.6.6",
     "accept": "^3.0.2",
     "apollo-server-core": "file:../apollo-server-core",
     "boom": "^7.1.0",
     "graphql-subscriptions": "^1.0.0",
-    "graphql-tools": "^4.0.0"
+    "graphql-tools": "^4.0.0",
+    "graphql-upload": "^8.0.0"
   },
   "devDependencies": {
     "apollo-server-integration-testsuite": "file:../apollo-server-integration-testsuite"
