+++ packages/apollo-server-koa/package.json
@@ -25,7 +25,6 @@
     "node": ">=6"
   },
   "dependencies": {
-    "@apollographql/apollo-upload-server": "^5.0.3",
     "@apollographql/graphql-playground-html": "^1.6.6",
     "@koa/cors": "^2.2.1",
     "@types/accepts": "^1.3.5",
@@ -38,6 +37,7 @@
     "apollo-server-core": "file:../apollo-server-core",
     "graphql-subscriptions": "^1.0.0",
     "graphql-tools": "^4.0.0",
+    "graphql-upload": "^8.0.0",
     "koa": "2.6.2",
     "koa-bodyparser": "^3.0.0",
     "koa-router": "^7.4.0",
