+++ packages/apollo-server-koa/src/ApolloServer.ts
@@ -12,7 +12,7 @@ import typeis from 'type-is';
 
 import { graphqlKoa } from './koaApollo';
 
-import { processRequest as processFileUploads } from '@apollographql/apollo-upload-server';
+import { processRequest as processFileUploads } from 'graphql-upload';
 
 export { GraphQLOptions, GraphQLExtension } from 'apollo-server-core';
 import { GraphQLOptions, FileUploadOptions } from 'apollo-server-core';
@@ -32,7 +32,11 @@ const fileUploadMiddleware = (
 ) => async (ctx: Koa.Context, next: Function) => {
   if (typeis(ctx.req, ['multipart/form-data'])) {
     try {
-      ctx.request.body = await processFileUploads(ctx.req, uploadsConfig);
+      ctx.request.body = await processFileUploads(
+        ctx.req,
+        ctx.res,
+        uploadsConfig,
+      );
       return next();
     } catch (error) {
       if (error.status && error.expose) ctx.status = error.status;
